+++ apps/meteor/client/views/room/modals/PinMessageModal/index.ts
@@ -1,0 +1,1 @@
+export { default } from './PinMessageModal';
